renamed to Envelope and fixed spacing
