--- spec/fixtures/components/foo-renderer/models/waf.rb
@@ -1,16 +1,0 @@
-module MyApp
-  class Waf < Alephant::Renderer::Views::Waf
-        def head
-          ['some/css/asset/path','some/js/asset/path']
-        end
-
-        def body_last
-          ['some/thing/to/add/after/body', 'some/other/thing/to/add/after/body']
-        end
-
-        def greet
-          @data[:foo]
-        end
-
-  end
-end
