+++ spec/fixtures/components/foo-renderer/templates/envelope.mustache
